Rearrange code for future extensions
To make it easy to add future extensions, the code is rearranged and
split into logical sections. This should give a clean view of where to
place new code.
