+++ netbox_inventory/api/serializers_/nested.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from netbox.api.serializers import WritableNestedSerializer
-from ..models import InventoryItemGroup
+from netbox_inventory.models import InventoryItemGroup
 
 __all__ = (
     'NestedInventoryItemGroupSerializer',
