+++ netbox_inventory/api/views.py
@@ -8,6 +8,30 @@ from .serializers import (
 )
 
 
+#
+# Assets
+#
+
+class InventoryItemGroupViewSet(NetBoxModelViewSet):
+    queryset = models.InventoryItemGroup.objects.add_related_count(
+        models.InventoryItemGroup.objects.all(),
+        models.Asset,
+        'inventoryitem_type__inventoryitem_group',
+        'asset_count',
+        cumulative=True
+    ).prefetch_related('tags')
+    serializer_class = InventoryItemGroupSerializer
+    filterset_class = filtersets.InventoryItemGroupFilterSet
+
+
+class InventoryItemTypeViewSet(NetBoxModelViewSet):
+    queryset = models.InventoryItemType.objects.prefetch_related('tags').annotate(
+        asset_count=count_related(models.Asset, 'inventoryitem_type')
+    )
+    serializer_class = InventoryItemTypeSerializer
+    filterset_class = filtersets.InventoryItemTypeFilterSet
+
+
 class AssetViewSet(NetBoxModelViewSet):
     queryset = models.Asset.objects.prefetch_related(
         'device_type', 'device', 'module_type', 'module', 'rack_type', 'rack',
@@ -16,6 +40,31 @@ class AssetViewSet(NetBoxModelViewSet):
     serializer_class = AssetSerializer
     filterset_class = filtersets.AssetFilterSet
 
+
+class DeviceAssetViewSet(DeviceViewSet):
+    """
+    Adds option to filter on asset assignemnet
+    """
+    filterset_class = filtersets.DeviceAssetFilterSet
+
+
+class ModuleAssetViewSet(ModuleViewSet):
+    """
+    Adds option to filter on asset assignemnet
+    """
+    filterset_class = filtersets.ModuleAssetFilterSet
+
+
+class InventoryItemAssetViewSet(InventoryItemViewSet):
+    """
+    Adds option to filter on asset assignemnet
+    """
+    filterset_class = filtersets.InventoryItemAssetFilterSet
+
+
+#
+# Deliveries
+#
 
 class SupplierViewSet(NetBoxModelViewSet):
     queryset = models.Supplier.objects.prefetch_related('tags').annotate(
@@ -42,38 +91,3 @@ class DeliveryViewSet(NetBoxModelViewSet):
     )
     serializer_class = DeliverySerializer
     filterset_class = filtersets.DeliveryFilterSet
-
-
-class InventoryItemTypeViewSet(NetBoxModelViewSet):
-    queryset = models.InventoryItemType.objects.prefetch_related('tags').annotate(
-        asset_count=count_related(models.Asset, 'inventoryitem_type')
-    )
-    serializer_class = InventoryItemTypeSerializer
-    filterset_class = filtersets.InventoryItemTypeFilterSet
-
-
-class InventoryItemGroupViewSet(NetBoxModelViewSet):
-    queryset = models.InventoryItemGroup.objects.add_related_count(
-        models.InventoryItemGroup.objects.all(),
-        models.Asset,
-        'inventoryitem_type__inventoryitem_group',
-        'asset_count',
-        cumulative=True
-    ).prefetch_related('tags')
-    serializer_class = InventoryItemGroupSerializer
-    filterset_class = filtersets.InventoryItemGroupFilterSet
-
-
-class DeviceAssetViewSet(DeviceViewSet):
-    """Adds option to filter on asset assignemnet"""
-    filterset_class = filtersets.DeviceAssetFilterSet
-
-
-class ModuleAssetViewSet(ModuleViewSet):
-    """Adds option to filter on asset assignemnet"""
-    filterset_class = filtersets.ModuleAssetFilterSet
-
-
-class InventoryItemAssetViewSet(InventoryItemViewSet):
-    """Adds option to filter on asset assignemnet"""
-    filterset_class = filtersets.InventoryItemAssetFilterSet
