+++ netbox_inventory/models/__init__.py
@@ -1,0 +1,2 @@
+from .assets import *
+from .deliveries import *
