+++ netbox_inventory/search.py
@@ -2,6 +2,27 @@ from netbox.search import SearchIndex
 from .models import Asset, Delivery, InventoryItemType, InventoryItemGroup, Purchase, Supplier
 
 
+#
+# Assets
+#
+
+class InventoryItemGroupIndex(SearchIndex):
+    model = InventoryItemGroup
+    fields = (
+        ("name", 100),
+        ("comments", 5000),
+    )
+
+
+class InventoryItemTypeIndex(SearchIndex):
+    model = InventoryItemType
+    fields = (
+        ("model", 100),
+        ("part_number", 100),
+        ("comments", 5000),
+    )
+
+
 class AssetIndex(SearchIndex):
     model = Asset
     fields = (
@@ -11,6 +32,11 @@ class AssetIndex(SearchIndex):
         ("comments", 5000),
     )
     display_attrs = ('name', 'asset_tag', 'status')
+
+
+#
+# Deliveries
+#
 
 class SupplierIndex(SearchIndex):
     model = Supplier
@@ -39,28 +65,11 @@ class DeliveryIndex(SearchIndex):
     )
 
 
-class InventoryItemTypeIndex(SearchIndex):
-    model = InventoryItemType
-    fields = (
-        ("model", 100),
-        ("part_number", 100),
-        ("comments", 5000),
-    )
-
-
-class InventoryItemGroupIndex(SearchIndex):
-    model = InventoryItemGroup
-    fields = (
-        ("name", 100),
-        ("comments", 5000),
-    )
-
-
 indexes = [
+    InventoryItemGroupIndex,
+    InventoryItemTypeIndex,
     AssetIndex,
     SupplierIndex,
     PurchaseIndex,
     DeliveryIndex,
-    InventoryItemTypeIndex,
-    InventoryItemGroupIndex,
 ]
